+++ src/components/CodeHighlight/CHANGELOG.md
@@ -2,3 +2,5 @@
 - 2025-09-26: Add support for TSX language in CodeHighlight.
 - 2025-10-05: Add `withLineNumbers` prop to CodeHighlight component.
 - 2025-10-11: Update line-numbers' style.
+- 2025-10-16: Refactor code to use hast to build the HTML. That make the
+  `CodeHighlight` component support `lineHighlight` prop, etc.